// src/payload-types.ts
export interface User {
  id: string;
  role: 'admin' | 'user';
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  _verified?: boolean | null;
  _verificationToken?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  password: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: string;
  user?: (string | null) | User;
  showFile?: boolean | null;
  cloudinary?: {
    public_id?: string | null;
    original_filename?: string | null;
    format?: string | null;
    secure_url?: string | null;
    resource_type?: string | null;
  };
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    card?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    tablet?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "digital-business-cards".
 */
export interface DigitalBusinessCard {
  id: string;
  userId: string | User;
  customName: string;
  fullName: string;
  jobTitle?: string | null;
  companyName?: string | null;
  companyLogo?: string | Media | null;
  profilePhoto?: string | Media | null;
  bannerImage?: string | Media | null;
  tagline?: string | null;
  bio?: string | null;
  about?: string | null;
  companyAddress?: string | null;
  googleMapsLink?: string | null;
  websiteURL?: string | null;
  emails?:
    | {
        title?: string | null;
        email?: string | null;
        id?: string | null;
      }[]
    | null;
  mobileNumbers?:
    | {
        title?: string | null;
        mobileNumber?: string | null;
        id?: string | null;
      }[]
    | null;
  whatsAppNumber?: string | null;
  customContactOptions?:
    | {
        linkName?: string | null;
        url?: string | null;
        icon?: string | Media | null;
        id?: string | null;
      }[]
    | null;
  socialMediaLinks?: {
    linkedin?: string | null;
    twitter?: string | null;
    instagram?: string | null;
    facebook?: string | null;
  };
  paymentInfo?: {
    upiLinks?:
      | {
          upiId?: string | null;
          id?: string | null;
        }[]
      | null;
    bank?: string | null;
    ifsc?: string | null;
    beneficiary?: string | null;
    accountNumber?: string | null;
    accountType?: string | null;
    gstNumber?: string | null;
  };
  qrCode?: string | Media | null;
  products?:
    | {
        name?: string | null;
        price?: number | null;
        discountPrice?: number | null;
        image?: string | Media | null;
        id?: string | null;
      }[]
    | null;
  photoGallery?:
    | {
        photo?: string | Media | null;
        id?: string | null;
      }[]
    | null;
  customSocialMediaLinks?:
    | {
        title?: string | null;
        url?: string | null;
        icon?: string | Media | null;
        id?: string | null;
      }[]
    | null;
  videos?:
    | {
        videoUrl?: string | null;
        id?: string | null;
      }[]
    | null;
  ctaButtons?:
    | {
        label?: string | null;
        url?: string | null;
        icon?: string | Media | null;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}